Finish implementing multi-process proxies
